--- src/css/fonts.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Open+Sans');
