+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from '../components/layout'
+import Layout from '../components/layout/layout'
 import Image from '../components/image'
 import SEO from '../components/seo'
 
@@ -9,8 +9,8 @@ const IndexPage = () => (
     <Layout>
         <SEO title="Home" />
         <h1>NDSU Cybersecurity Student Association Homepage</h1>
-        <p>We are a club of many individuals that share a common interest of creating technology, and breaking it.</p>
-        <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
+        <p>We are a group of individuals that share a common interest of creating technology, and breaking it.</p>
+        <div style={{ maxWidth: `300px`, marginBottom: `1.45rem`, margin: `0 auto`}}>
             <Image />
         </div>
     </Layout>
