+++ src/pages/meetings.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Layout from '../components/layout'
+import Layout from '../components/layout/layout'
 import SEO from '../components/seo'
 import { StaticQuery, graphql } from 'gatsby';
